# Do not encode message to check its validity after decoding