@guelfey 2023-04-09
Included in https://github.com/godbus/dbus/pull/364. Thanks!